app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
SQLAlchemy>=2.0
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
passlib>=1.7
edge-tts>=6.1